remove sqlite3 from test group
Since it is already in the `:development, :test` block, it doesn't need
to be specified again in the test group.

Otherwise bundler gives me this warning:
```
Your Gemfile lists the gem sqlite3 (>= 0) more than once.
You should probably keep only one of them.
While it's not a problem now, it could cause errors if you change the
version of just one of them later.
```